+++ vtm-android-example/src/org/oscim/android/test/MBTilesBitmapActivity.java
@@ -37,7 +37,7 @@ public class MBTilesBitmapActivity extends BitmapTileActivity {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        File file = new File(getExternalFilesDir(null), "raster.mbtiles");
+        File file = new File(getExternalMediaDirs()[0], "raster.mbtiles");
         if (!file.exists()) {
             AlertDialog.Builder builder = new AlertDialog.Builder(this)
                     .setTitle(R.string.warning)