+++ docs/Changelog.md
@@ -6,10 +6,11 @@
   - Removed `pointReduction` option
 - Line style `transparent` option [#1100](https://github.com/mapsforge/vtm/pull/1100)
 - Canvas adapter: line scale option [#1075](https://github.com/mapsforge/vtm/pull/1075)
+- Disable `vtm-app`
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.21.0)
 
-## Version 0.19.0 (2023-08-15)
+## Version 0.20.0 (2023-08-15)
 
 - Motorider map theme [#1045](https://github.com/mapsforge/vtm/pull/1045)
 - Minor improvements and bug fixes